requirements.txt
azure-functions
redis
requests